requirements.txt
anyio==4.4.0
asttokens==2.4.1
boto3==1.34.144
botocore==1.34.144
catboost==1.2.5
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
evidently==0.2.8
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.1
from-root==1.3.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-boto3-s3==1.34.138
nest-asyncio==1.6.0
neuro-mf==0.0.5
nltk==3.8.1
numpy==2.0.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pydantic==1.10.17
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
tenacity==8.5.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
-e git+https://github.com/utkarsh-iitbhu/US-Visa-automate.git@28d933d6e1d7564ea47f0ded1f58cad5da228471#egg=us_visa
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
xgboost==2.1.0
